+++ extensions/premier-league/src/components/searchbar_season.tsx
@@ -1,21 +1,35 @@
 import { List } from "@raycast/api";
 import { usePromise } from "@raycast/utils";
+import { useState } from "react";
 import { getSeasons } from "../api";
+import { Season } from "../types";
+
+export const useSeasonSelection = () => {
+  const [selectedSeasonId, setSelectedSeasonId] = useState<string>();
+  const { data: seasons = [], isLoading } = usePromise(getSeasons);
+
+  return {
+    seasonId: selectedSeasonId ?? seasons[0]?.seasonId,
+    setSeasonId: setSelectedSeasonId,
+    seasons,
+    isLoading,
+  };
+};
 
 export default function SearchBarSeason(props: {
   selected?: string;
   onSelect: React.Dispatch<React.SetStateAction<string | undefined>>;
+  seasons: Season[];
+  isLoading: boolean;
 }) {
-  const { data: seasons, isLoading } = usePromise(getSeasons);
-
   return (
     <List.Dropdown
       tooltip="Filter by Season"
       value={props.selected}
       onChange={props.onSelect}
-      isLoading={isLoading}
+      isLoading={props.isLoading}
     >
-      {seasons?.map((season) => {
+      {props.seasons.map((season) => {
         return (
           <List.Dropdown.Item
             key={season.seasonId}